+++ lib/storage_hub.dart
@@ -16,20 +16,51 @@ class StorageHub {
   StorageHub._();
   static final StorageHub hub = StorageHub._();
 
+  /// returns `true` if configuration completed
   static bool isConfigured = false;
+
+  /// Base url of the API end point of your Storage Hub
   static String apiEndpointBaseUrl = '';
+
+  /// Part url of your upload request to obtain a session id
   static String apiEndpointPostUrl = '?uploadType=resumable&name=';
+
+  /// Part url of your upload operation
   static String apiEndpointPutUrl = '?upload_id=';
+
+  /// api key of your storage hub
   static String apiSecurityKey = '';
+
+  /// the length of file parts to send in one chunk.
   static int chunkSizeInBytes = 1024 * 89;
+
+  /// retry count to determine that operation is actually an error.
+  /// when an upload task returns error, `errorCount` is increased
+  /// by one and `processStartTime` updates 15 minutes from current time.
+  /// The error might be temporary, so 15 minutes later it will retry
+  /// the same file. if it fails again, it will increase `errorCount` and
+  /// queue for another 15 minutes.
+  /// After `errorTreshold` times the file will be marked as error and wont
+  /// retry anymore.
   static int errorTreshold = 100;
 
-  static List<FileModel> fileList = [];
+  /// list of your files
+  static List<FileDefinition> fileList = [];
+
+  /// returns `true` if any file is uploading currently
   static bool isSyncing = false;
-  static FileModel? syncingFile;
+
+  /// returns `FileDefinition` of the file that is
+  /// currently being uploaded.
+  /// It will return null if upload operation is idle.
+  static FileDefinition? syncingFile;
+
+  /// progress of the current upload operation. min:0, max:1
+  /// it will return 0 if there is no operation.
   static double progress = 0;
 
-  configure({
+  /// Configuration/Initialization method
+  static configure({
     required String baseUrl,
     required String apiKey,
     String? postUrl = '?uploadType=resumable&name=',
@@ -46,6 +77,13 @@ class StorageHub {
     isConfigured = true;
   }
 
+  /// add a file to database
+  /// filePath, fileName, totalBytes are required
+  /// time is optional,
+  /// `DateTime.now().toIso8601String()` will be used if null
+  /// metadata is a map that you may want to store any
+  /// info related to your file.
+  /// it will be sent to Storage Hub during upload.
   static Future<bool> addFile({
     required String filePath,
     required String fileName,
@@ -53,7 +91,7 @@ class StorageHub {
     Map<String, dynamic>? metadata,
     String? time,
   }) async {
-    FileModel file = FileModel(
+    FileDefinition file = FileDefinition(
       filePath: filePath,
       fileName: fileName,
       time: time ?? DateTime.now().toIso8601String(),
@@ -69,6 +107,19 @@ class StorageHub {
     return result > 0;
   }
 
+  /// Deletes the file specified by id
+  /// id can be obtained from `FileDefinition file.id`
+  /// from your list
+  static Future<bool> deleteFile({required int id}) async {
+    int result = await _DbProvider.db.deleteFile(id: id);
+    fileList = await _DbProvider.db.getFilesByStatus();
+    return result > 0;
+  }
+
+  /// Triggers synchronozation of files that are waiting for upload
+  /// The operation will perform for the first file in queue depending
+  /// on status of the file.
+  /// If the file did started to upload before, this time it will resume
   static void triggerSync() async {
     if (isSyncing) return;
     isSyncing = true;
@@ -77,10 +128,18 @@ class StorageHub {
     await _startProgress();
   }
 
-  static Future<FileModel?> _populateFileList() async {
+  /// This will return a fileList
+  /// Different than `StorageHub.fileList`, this will
+  /// use SQLite data source instead of ram.
+  static Future<List<FileDefinition>> getFileList() async {
+    return await _DbProvider.db.getFilesByStatus();
+  }
+
+  static Future<FileDefinition?> _populateFileList() async {
     fileList = await _DbProvider.db.getFilesByStatus();
     if (fileList.isEmpty) return null;
-    fileList.sort((a, b) => a.processStartTime ?? 0.compareTo(b.processStartTime ?? 0));
+    fileList.sort(
+        (a, b) => a.processStartTime ?? 0.compareTo(b.processStartTime ?? 0));
     var list = fileList.where((e) => e.status == SyncStatus.idle).toList();
     return list.isNotEmpty ? list.first : null;
   }
@@ -122,7 +181,8 @@ class StorageHub {
           await _DbProvider.db.deleteFile(id: syncingFile!.id!);
         } else {
           // set file to retry again
-          syncingFile!.processStartTime = DateTime.now().millisecondsSinceEpoch + (1000 * 60 * 15);
+          syncingFile!.processStartTime =
+              DateTime.now().millisecondsSinceEpoch + (1000 * 60 * 15);
           syncingFile!.status = SyncStatus.idle;
           await _DbProvider.db.updateFile(file: syncingFile!);
         }
@@ -141,6 +201,12 @@ class StorageHub {
   }
 }
 
+/// Enum for Synchronization States
+/// idle: file is queued for upload
+/// requestingUpload: file has a session for upload task
+/// uploading: file is currently being uploading
+/// uploaded: file has been uploaded
+/// error: file upload task had encountered an error
 enum SyncStatus {
   idle,
   requestingUpload,
@@ -149,7 +215,7 @@ enum SyncStatus {
   error,
 }
 
-class FileModel {
+class FileDefinition {
   int? id;
   String? time;
   String filePath;
@@ -161,7 +227,7 @@ class FileModel {
   int errorCount;
   int? processStartTime;
   Map<String, dynamic> metadata;
-  FileModel({
+  FileDefinition({
     this.id,
     this.time,
     required this.filePath,
@@ -202,7 +268,7 @@ class FileModel {
         'metadata': metadata,
       };
 
-  factory FileModel.fromMap(Map<String, dynamic> map) => FileModel(
+  factory FileDefinition.fromMap(Map<String, dynamic> map) => FileDefinition(
         id: map['id']?.toInt(),
         time: map['time'] ?? '',
         filePath: map['filePath'] ?? '',
@@ -253,7 +319,7 @@ class _DbProvider {
     // TODO: Restore data
   }
 
-  Future<int> insertFile({required FileModel file}) async {
+  Future<int> insertFile({required FileDefinition file}) async {
     final db = await database;
     if (db == null) return -1;
     try {
@@ -285,7 +351,7 @@ class _DbProvider {
     }
   }
 
-  Future<int> updateFile({required FileModel file}) async {
+  Future<int> updateFile({required FileDefinition file}) async {
     final db = await database;
     if (db == null) return -1;
     try {
@@ -303,7 +369,7 @@ class _DbProvider {
     }
   }
 
-  Future<FileModel?> getFileById({required int id}) async {
+  Future<FileDefinition?> getFileById({required int id}) async {
     final db = await database;
     if (db == null) return null;
     try {
@@ -315,15 +381,15 @@ class _DbProvider {
         orderBy: _Keys.id,
         limit: 1,
       );
-      if (result.isNotEmpty) return FileModel.fromMap(result[0]);
+      if (result.isNotEmpty) return FileDefinition.fromMap(result[0]);
     } catch (err) {
       log(err.toString(), name: tag);
     }
     return null;
   }
 
-  Future<List<FileModel>> getFilesByStatus({SyncStatus? status}) async {
-    List<FileModel> fileList = [];
+  Future<List<FileDefinition>> getFilesByStatus({SyncStatus? status}) async {
+    List<FileDefinition> fileList = [];
     final db = await database;
     if (db == null) return fileList;
     try {
@@ -337,7 +403,7 @@ class _DbProvider {
       );
       if (result.isNotEmpty) {
         for (var row in result) {
-          fileList.add(FileModel.fromMap(row));
+          fileList.add(FileDefinition.fromMap(row));
         }
       }
     } catch (err) {
@@ -353,14 +419,17 @@ class _NetworkProvider {
     if (headers != null) dio.options.headers.addAll(headers);
     if (contentType != null) dio.options.contentType = contentType;
     dio.options.followRedirects = false;
-    (dio.httpClientAdapter as DefaultHttpClientAdapter).onHttpClientCreate = (HttpClient dioClient) {
-      dioClient.badCertificateCallback = ((X509Certificate cert, String host, int port) => true);
+    (dio.httpClientAdapter as DefaultHttpClientAdapter).onHttpClientCreate =
+        (HttpClient dioClient) {
+      dioClient.badCertificateCallback =
+          ((X509Certificate cert, String host, int port) => true);
       return dioClient;
     };
     return dio;
   }
 
-  static Future<FileModel> requestUpload({required FileModel file}) async {
+  static Future<FileDefinition> requestUpload(
+      {required FileDefinition file}) async {
     Map<String, dynamic> request = file.metadata;
     request['time'] = file.time;
     request['fileName'] = file.fileName;
@@ -373,10 +442,14 @@ class _NetworkProvider {
     };
 
     Dio dio = _prepareDio(headers: headers, contentType: "application/json");
-    String url = "${StorageHub.apiEndpointBaseUrl}${StorageHub.apiEndpointPostUrl}${file.fileName}";
+    String url =
+        "${StorageHub.apiEndpointBaseUrl}${StorageHub.apiEndpointPostUrl}${file.fileName}";
     Response response = await dio.post(url, data: request);
-    if (response.statusCode == HttpStatus.ok || response.statusCode == HttpStatus.created) {
-      if (response.data != null && response.data['success'] == true && response.data['statusCode'] == HttpStatus.ok) {
+    if (response.statusCode == HttpStatus.ok ||
+        response.statusCode == HttpStatus.created) {
+      if (response.data != null &&
+          response.data['success'] == true &&
+          response.data['statusCode'] == HttpStatus.ok) {
         String sessionId = response.data['data'][0]['id'];
         file.sessionId = sessionId;
         file.status = SyncStatus.idle;
@@ -392,7 +465,8 @@ class _NetworkProvider {
     return file;
   }
 
-  static Future<FileModel> putFile({required FileModel file, Function(int, int)? progress}) async {
+  static Future<FileDefinition> putFile(
+      {required FileDefinition file, Function(int, int)? progress}) async {
     //..
     String contentType = _Utils.getContentType(file.fileName);
     Map<String, dynamic> headers = {
@@ -401,15 +475,19 @@ class _NetworkProvider {
       "Content-Range": "bytes */*",
       "Accept-Encoding": "gzip,deflate,br",
       "Accept": "*/*",
-      "Content-Length": math.min(StorageHub.chunkSizeInBytes, file.totalBytes - file.uploadedBytes),
+      "Content-Length": math.min(
+          StorageHub.chunkSizeInBytes, file.totalBytes - file.uploadedBytes),
     };
 
     Dio dio = _prepareDio(headers: headers);
-    String url = "${StorageHub.apiEndpointBaseUrl}${StorageHub.apiEndpointPutUrl}${file.sessionId}";
+    String url =
+        "${StorageHub.apiEndpointBaseUrl}${StorageHub.apiEndpointPutUrl}${file.sessionId}";
 
     final int startByte = file.uploadedBytes;
-    final int endByte = math.min(startByte + StorageHub.chunkSizeInBytes, file.totalBytes);
-    final Stream<List<int>> chunkStream = File(file.filePath).openRead(startByte, endByte);
+    final int endByte =
+        math.min(startByte + StorageHub.chunkSizeInBytes, file.totalBytes);
+    final Stream<List<int>> chunkStream =
+        File(file.filePath).openRead(startByte, endByte);
 
     Response? r;
     try {
@@ -461,7 +539,8 @@ class _NetworkProvider {
 class _Constants {
   static const String databaseName = 'storageHubDb.sqlite';
   static const int databaseVersion = 1;
-  static const String dropStorageTable = 'DROP TABLE IF EXISTS ${_Keys.tableName}';
+  static const String dropStorageTable =
+      'DROP TABLE IF EXISTS ${_Keys.tableName}';
   static const String createStorageTable = '''
     CREATE TABLE IF NOT EXISTS ${_Keys.tableName} (
       ${_Keys.id} INTEGER PRIMARY KEY AUTOINCREMENT,