+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.2
+* Added `deleteFile` method
+```
+bool deleted = await StorageHub.deleteFile(id: 123);
+```
+* Renamed `FileModel` to `FileDefinition`
+* Added `getFileList` method for fetching file list from sqlite
+```
+List<FileDefinition> files = await StorageHub.getFileList();
+```
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial Release